requirements.txt
numpy
matplotlib
scipy
xarray
netCDF4
pandas
pyyaml
torch
